+++ src/corelib/hall-conf-dfl.h
@@ -18,13 +18,30 @@
 #define HALL_SPEED_ENABLED  1
 #endif
 
-#define HALL_SWITH_DEVICE       
-
-#define HALL_SWITCH_PLATFORM    HALL_SWITCH_WICED
-
 #define HALL_SWITCH_FRAMEWORK_ARDUINO   0x00U
 #define HALL_SWITCH_FRAMEWORK_WICED     0x01U
 
-#define HALL_SWITCH_FRAMEWORK   HALL_SWITCH_FRAMEWORK_ARDUINO
+#ifndef HALL_SWITCH_FRAMEWORK 
+#error  "Hall switch error framework undefined"
+#define HALL_SWITCH_FRAMEWORK 
+#endif
+
+#ifndef HALL_SWITCH_PLATFORM    
+#define HALL_SWITCH_PLATFORM    HALL_SWITCH_PLATF_ALL
+#endif
+
+//#define HALL_SWITCH_FRAMEWORK   HALL_SWITCH_FRAMEWORK_ARDUINO
+#define HALL_SWITCH_MCU_BOARD           
+#define XMC2GO
+#define UNO_PLUS_MY_IOT_ADAPTER_SOCKET23
+
+#define HALL_SWITCH_SENSOR_SHIELD
+
+#define SPEED_SENSOR_2GO_KIT_TLE4922
+
+#define HALL_SWITCH_PLATF_ARD_XM2GO_TLE     0x00U
+#define HALL_SWITCH_PLATF_ARD_UNO_          0x
+#define HALL_SWITCH_PLATF_KIT2GO_SPEED_
+#define HALL_SWITCH_PLAT_WICED_
 
 #endif /** HALL_CONF_DFL_H_ **/
