+++ src/corelib/hall-switch.h
@@ -3,6 +3,7 @@
  * @brief       Hall Switch Hardware Abstraction Layer
  * @date        July 2019
  * @copyright   Copyright (c) 2019-2020 Infineon Technologies AG
+ * SPDX-License-Identifier: MIT
  */
 
 #ifndef HALL_SWITCH_H_
