+++ src/framework/arduino/pal/hall-pal-ino.h
