+++ src/framework/wiced-43xxx/pal/hall-pal-wiced.cpp
