+++ src/framework/wiced-43xxx/wrapper/hall-platf-wiced.cpp
