+++ src/framework/wiced-43xxx/wrapper/hall-switch-wiced.h
